separation of powers
basically the python script was getting annoying to manage so i just split up the functions into different files which makes it easier to see exactly what is where. i think this is better. next thing will be modifying the runner files to match
