+++ app/sleep.py
@@ -1,0 +1,5 @@
+import time
+
+# sleep method, makes main code a little more readable
+def sleep(n):
+	time.sleep (n)
